+++ R/guides.R
@@ -55,8 +55,8 @@ guide_ggchord_colourbar <- function(
     (!is.null(position) && position %in% c("top", "bottom"))
   compact <- ggplot2::theme(
     legend.title.position = "top",
-    legend.key.width = unit(if (horizontal) 42 else 3, "mm"),
-    legend.key.height = unit(if (horizontal) 3 else 42, "mm"),
+    legend.key.width = unit(if (horizontal) 46 else 3, "mm"),
+    legend.key.height = unit(if (horizontal) 3 else 46, "mm"),
     legend.ticks.length = unit(1, "mm"),
     legend.text = element_text(margin = margin(l = 1.2)),
     legend.title = element_text(margin = margin(b = 1.5))